+++ CHANGELOG.md
@@ -9,11 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enable all CORS requests.
 - Airtable bulk actions records chunking.
 - Cart customer details save feature.
-- Stripe payment intent customer creation.
+- Stripe Payment Intent customer creation.
 - Subtotal free shipping.
 
 ### Changed
 - Rename project.
+
+### Fixed
+- Update Stripe Payment Intent shipping amount metadata when cart gets updated.
 
 ### Removed
 - Remove time fields in cart to order items mapping.